cleanup a bit; test still fails

+++ src/Enclave/external/keccak.h
@@ -20,7 +20,7 @@ extern "C" {
 #endif
 
 // compute a keccak hash (md) of given byte length from "in"
-int keccak(const uint8_t *in, int inlen, uint8_t *md, int mdlen);
+int keccak(const uint8_t *in, size_t inlen, uint8_t *md, size_t mdlen);
 
 // update the state
 void keccakf(uint64_t st[25], int norounds);
